requirements.txt
click
gdal
mappyfile
pyyaml